Refactor: move classes to a dedicated file
