+++ src/heating_tempflow_iotedge/modules/AddTemperatureCorrectionFactor/Program.cs
@@ -142,33 +142,4 @@ namespace AddTemperatureCorrectionFactor
             await client.UpdateReportedPropertiesAsync(reportedProperties);
         }
     }
-
-    public class SensorData
-    {
-        [JsonProperty("sensorid")]
-        public string SensorId { get; set; }
-
-        [JsonProperty("timestamp")]
-        public DateTime MeasurementDateTime { get; set; }
-
-        [JsonProperty("temperature")]
-        public double Temperature { get; set; }
-    }
-
-    public class SensorCorrectionFactor
-    {
-        public string SensorId { get; set; }
-        public string SensorDescription { get; set; }
-        public float CorrectionFactor { get; set; }
-
-        public static SensorCorrectionFactor DefaultFor(string sensorId)
-        {
-            return new SensorCorrectionFactor
-            {
-                SensorId = sensorId,
-                    SensorDescription = "Unknown",
-                    CorrectionFactor = 0.0f
-            };
-        }
-    }
 }
