+++ src/heating_tempflow_iotedge/modules/AddTemperatureCorrectionFactor/SensorData.cs
@@ -1,0 +1,17 @@
+namespace AddTemperatureCorrectionFactor
+{
+    using System;
+    using Newtonsoft.Json;
+
+    public class SensorData
+    {
+        [JsonProperty("sensorid")]
+        public string SensorId { get; set; }
+
+        [JsonProperty("timestamp")]
+        public DateTime MeasurementDateTime { get; set; }
+
+        [JsonProperty("temperature")]
+        public double Temperature { get; set; }
+    }
+}
